Web UI: move queue management under 'Status' section
