+++ root/usr/share/nethesis/NethServer/Language/en/NethServer_Module_Mail.php
@@ -31,7 +31,7 @@ $L['TransportType_Reject_label'] = 'Reject';
 $L['TransportType_Relay_label'] = 'Relay to another server'; 
 $L['Update domain `${0}`'] = 'Update domain `${0}`'; 
 $L['domain_label'] = 'Domain';
-$L['Queue_Title'] = 'Queue management';
+$L['MailQueue_Title'] = 'Mail queue';
 $L['AndXMore'] = '..and ${0} more';
 $L['Refresh_label'] = 'Refresh';
 $L['Id_label'] = 'Id';
