+++ root/usr/share/nethesis/NethServer/Module/Mail/MailQueue/Refresh.php
@@ -1,5 +1,5 @@
 <?php
-namespace NethServer\Module\Mail\Queue;
+namespace NethServer\Module\MailQueue;
 
 /*
  * Copyright (C) 2012 Nethesis S.r.l.
@@ -37,7 +37,6 @@ class Refresh extends \Nethgui\Controller\Table\AbstractAction
 
     public function nextPath()
     {
-        return '/Mail/Queue/read';
+        return '/MailQueue/read';
     }
-
 }
