+++ root/usr/share/nethesis/NethServer/Template/MailQueue/Delete.php
