+++ root/usr/share/nethesis/NethServer/Template/MailQueue/DeleteAll.php
